[#185] 모델 변경사항 반영
### Description
- 모델 변경사항 반영
- 더미데이터 수정

### Related Issue
Resolve #185
