+++ Journey/Journey/Sources/APIModels/Challenge.swift
@@ -10,22 +10,23 @@ import Foundation
 // MARK: - Challenge
 struct Challenge: Codable {
     let id, situation: Int
-    let title, challengeDescription, year, month: String
-    let day: String
-    let currentStamp, totalStamp: Int
-    let userMents: [String]
+        let title, challengeDescription, successDescription, year: String
+        let month, day: String
+        let currentStamp, totalStamp: Int
+        let userMents: [String]
 
-    enum CodingKeys: String, CodingKey {
-        case id, situation, title
-        case challengeDescription = "description"
-        case year, month, day, currentStamp, totalStamp, userMents
-    }
+        enum CodingKeys: String, CodingKey {
+            case id, situation, title
+            case challengeDescription = "description"
+            case successDescription, year, month, day, currentStamp, totalStamp, userMents
+        }
     
-    init(id: Int, situation: Int, title: String, challengeDescription: String, year: String, month: String, day: String, currentStamp: Int, totalStamp: Int, userMents: [String]) {
+    init(id: Int, situation: Int, title: String, challengeDescription: String, successDescription: String, year: String, month: String, day: String, currentStamp: Int, totalStamp: Int, userMents: [String]) {
         self.id = id
         self.situation = situation
         self.title = title
         self.challengeDescription = challengeDescription
+        self.successDescription = successDescription
         self.year = year
         self.month = month
         self.day = day
