+++ Journey/Journey/Sources/APIModels/Course.swift
@@ -34,7 +34,7 @@ struct Course: Codable {
     let id, situation: Int
     let title, courseDescription: String
     let totalDays: Int
-    let property: String
+    let property: Int
     let challenges: [Challenge]
 
     enum CodingKeys: String, CodingKey {
@@ -43,7 +43,7 @@ struct Course: Codable {
         case totalDays, property, challenges
     }
     
-    init(id: Int, title: String, courseDescription: String, totalDays: Int, situation: Int, property: String, challenges: [Challenge]) {
+    init(id: Int, title: String, courseDescription: String, totalDays: Int, situation: Int, property: Int, challenges: [Challenge]) {
         self.id = id
         self.situation = situation
         self.title = title
