+++ Journey/Journey/Sources/Cells/CourseLibraryCollectionVIew/DoneCourseCollectionViewCell.swift
@@ -14,7 +14,7 @@ class DoneCourseCollectionViewCell: UICollectionViewCell {
     var courseViewModel: CourseViewModel! {
         didSet {
             titleLabel.text = courseViewModel.course.title
-            coursePropertyLabel.text = courseViewModel.course.property
+            coursePropertyLabel.text = "\(courseViewModel.course.property)"
             courseDaysLabel.text = "\(courseViewModel.course.totalDays)일"
             descriptionTextView.text = courseViewModel.course.courseDescription
