+++ Journey/Journey/Sources/ViewControllers/Challenge/CourseLibraryViewController.swift
@@ -11,15 +11,7 @@ class CourseLibraryViewController: UIViewController {
     
     // MARK: - Properties
     // dummy data
-    var courses: [Course] = [
-        Course(id: 1, title: "뽀득뽀득 세균 퇴치", courseDescription: "나 쟈니가 인간세계에 처음 도착했을 때 사람들이 청결에 대해 은근히 무심한 것이 신기했쟈니. 내가 사는 별에서는 상상도 할 수 없쟈니.", totalDays: 7, situation: 0, property: "water", challenges: []),
-        Course(id: 2, title: "가나다라마바사 퇴치", courseDescription: "해당 합의안을 중대본에 전달하고, 최종 결과는 이날 오후 발표할 예정이다. 해당 합의안을 중대본에 전달하고, 최종", totalDays: 2, situation: 0, property: "water", challenges: []),
-        Course(id: 3, title: "쟈니 인간세계 오기", courseDescription: "얘들아 오늘 일 하다가 심심하면 마니또 삼행시 이벤트 참여해줘,,,얘들아 오늘 일 하다가 심심하면 마니또 삼행시 이벤트 참여해줘,,,", totalDays: 10, situation: 0, property: "water", challenges: []),
-        Course(id: 4, title: "이것은 더미데이터", courseDescription: "걍 아무글이나 써보는 중인데 이렇게 짧으면 어케 나오나", totalDays: 7, situation: 0, property: "water", challenges: []),
-        Course(id: 6, title: "더미데이터 입니다", courseDescription: "그색긴널사랑하는게아냐 ~그색긴널사랑하는게아냐 ~그색긴널사랑하는게아냐 ~그색긴널사랑하는게아냐 ~그색긴널사랑하는게아냐 ~", totalDays: 4, situation: 2, property: "water", challenges: []),
-        Course(id: 7, title: "이미 한 것들 잘 나오나", courseDescription: "델타변이 가주앙~!!!델타변이 가주앙~!!!델타변이 가주앙~!!!델타변이 가주앙~!!!", totalDays: 7, situation: 2, property: "water", challenges: []),
-        Course(id: 8, title: "어제 끝난 코스임", courseDescription: "마지막 셀입니다요. 마지막 셀입니다요. 마지막 셀입니다요. 마지막 셀입니다요. 마지막 셀입니다요. 마지막 셀입니다요. ", totalDays: 5, situation: 2, property: "water", challenges: [])
-    ]
+    var courses: [Course] = []
     
     var doingCourse: Bool = true
     var courseListViewModel = CourseListViewModel()
