+++ Journey/Journey/Sources/ViewControllers/Challenge/CourseViewController.swift
@@ -13,9 +13,9 @@ class CourseViewController: UIViewController {
     // MARK: - Properties
 
     // default data
-    var course: Course = Course(id: 0, title: "", courseDescription: "", totalDays: 0, situation: 0, property: "", challenges: [
-        Challenge(id: 0, situation: 0, title: "", challengeDescription: "", year: "", month: "", day: "", currentStamp: 0, totalStamp: 0, userMents: []),
-        Challenge(id: 0, situation: 0, title: "", challengeDescription: "", year: "", month: "", day: "", currentStamp: 0, totalStamp: 0, userMents: []),
+    var course: Course = Course(id: 0, title: "", courseDescription: "", totalDays: 0, situation: 0, property: 0, challenges: [
+        Challenge(id: 0, situation: 0, title: "", challengeDescription: "", successDescription: "", year: "", month: "", day: "", currentStamp: 0, totalStamp: 0, userMents: []),
+        Challenge(id: 0, situation: 0, title: "", challengeDescription: "", successDescription: "", year: "", month: "", day: "", currentStamp: 0, totalStamp: 0, userMents: []),
     ])
     
     // MARK: - @IBOutlet Properties
